+++ earthpy/tests/conftest.py
@@ -1,128 +1,11 @@
 """ Utility functions for tests. """
 import numpy as np
 import pytest
-from shapely.geometry import Polygon, Point, LineString
 from affine import Affine
 import rasterio as rio
-import geopandas as gpd
 
 
-def make_locs_gdf():
-    """ Create a dummy point GeoDataFrame. """
-    pts = np.array([[2, 2], [3, 4], [9, 8], [-12, -15]])
-    gdf = gpd.GeoDataFrame(
-        [Point(xy) for xy in pts],
-        columns=["geometry"],
-        crs={"init": "epsg:4326"},
-    )
-    return gdf
-
-
-def make_poly_in_gdf():
-    """ Bounding box polygon. """
-    poly_inters = Polygon([(0, 0), (0, 10), (10, 10), (10, 0), (0, 0)])
-    gdf = gpd.GeoDataFrame(
-        [1], geometry=[poly_inters], crs={"init": "epsg:4326"}
-    )
-    return gdf
-
-
-def make_locs_buff():
-    """ Buffer points to create multi poly. """
-    buffered_locations = make_locs_gdf()
-    buffered_locations["geometry"] = buffered_locations.buffer(4)
-    return buffered_locations
-
-
-def make_donut_geom():
-    """ Make a donut geometry. """
-    donut = gpd.overlay(
-        make_locs_buff(), make_poly_in_gdf(), how="symmetric_difference"
-    )
-    return donut
-
-
-@pytest.fixture
-def locs_gdf():
-    """ Get a dummy point GeoDataFrame.
-
-    This fixture calls make_locs_gdf(), which is a function that is used in
-    multiple fixtures. But, fixtures are not supposed to be used like that:
-
-    see https://github.com/pytest-dev/pytest/issues/3950 for discussion
-    """
-    return make_locs_gdf()
-
-
-@pytest.fixture
-def linez_gdf():
-    """ Create Line Objects For Testing """
-    linea = LineString([(1, 1), (2, 2), (3, 2), (5, 3)])
-    lineb = LineString([(3, 4), (5, 7), (12, 2), (10, 5), (9, 7.5)])
-    gdf = gpd.GeoDataFrame(
-        [1, 2], geometry=[linea, lineb], crs={"init": "epsg:4326"}
-    )
-    return gdf
-
-
-@pytest.fixture
-def poly_in_gdf():
-    """ Fixture for a bounding box polygon. """
-    return make_poly_in_gdf()
-
-
-@pytest.fixture
-def locs_buff():
-    """ Fixture for buffered locations. """
-    return make_locs_buff()
-
-
-@pytest.fixture
-def donut_geom():
-    """ Fixture for donut geometry objects. """
-    return make_donut_geom()
-
-
-@pytest.fixture
-def multi_gdf():
-    """ Create a multi-polygon GeoDataFrame. """
-    multi_poly = make_donut_geom().unary_union
-    out_df = gpd.GeoDataFrame(
-        gpd.GeoSeries(multi_poly), crs={"init": "epsg:4326"}
-    )
-    out_df = out_df.rename(columns={0: "geometry"}).set_geometry("geometry")
-    return out_df
-
-
-@pytest.fixture
-def basic_geometry():
-    """
-    A square polygon spanning (2, 2) to (4.25, 4.25) in x and y directions
-    Borrowed from rasterio/tests/conftest.py
-
-    Returns
-    -------
-    dict: GeoJSON-style geometry object.
-        Coordinates are in grid coordinates (Affine.identity()).
-    """
-    return Polygon([(2, 2), (2, 4.25), (4.25, 4.25), (4.25, 2), (2, 2)])
-
-
-@pytest.fixture
-def basic_geometry_gdf(basic_geometry):
-    """
-    A GeoDataFrame containing the basic geometry
-
-    Returns
-    -------
-    GeoDataFrame containing the basic_geometry polygon
-    """
-    gdf = gpd.GeoDataFrame(
-        geometry=[basic_geometry], crs={"init": "epsg:4326"}
-    )
-    return gdf
-
-
+# used several times
 @pytest.fixture
 def basic_image():
     """
@@ -139,6 +22,7 @@ def basic_image():
     return image
 
 
+# used several times
 @pytest.fixture
 def basic_image_tif(tmpdir, basic_image):
     """
@@ -165,15 +49,7 @@ def basic_image_tif(tmpdir, basic_image):
     return outfilename
 
 
+# used several times
 @pytest.fixture
 def image_array_2bands():
     return np.random.randint(10, size=(2, 4, 5))
-
-
-@pytest.fixture
-def one_band_3dims():
-    """Return a 3-dim numpy array vals 0-9"""
-
-    return np.array(
-        [[[8, 0, 2, 6, 3], [2, 8, 2, 8, 4], [3, 9, 1, 5, 4], [5, 9, 2, 7, 7]]]
-    )
